+++ CHANGELOG.md
@@ -8,11 +8,23 @@ Changed
 
 Added
 
+
+[18.6.4] - 2018-06-19
+Added
+    test_calculate_ims.py
+Changed
+    modified calculate_im script with added -s --simple_output option to only output summary csv
+
+
 [18.6.3] - 2018-06-14
 Added
     Slurm script generation to aide submission for the job to Kupe/Maui/Mahuika for binary and ascii workflows
     aggregation of IMs across realisations script
     Option to select acc units for each run
+Changed
+    modified calculate_im script to use unique ascending periods when concatenating with extended periods
+    modified calculate_im script to use sorted stations when writing result csvs
+
 
 [18.6.2] - 2018-06-11
 Added
@@ -36,8 +48,7 @@ Added
     rrup calculation script
     R_x template code
 Changed
-    modified rrup.csv output to conform to the template on wiki	
-
+    modified rrup.csv output to conform to the template on wiki